% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotations_dictionary}
\alias{annotations_dictionary}
\title{Annotation dictionary to keep track of annotations changes and versions}
\format{
A list of named vectors with the correspondence between time-stamped
annotations
}
\source{
This file was created with the script in
"inst/scripts/make-annotation-dictionary.R
}
\description{
Annotations are opinionated and dynamic by nature. They are subjected to
change as more experts look into the data and more evidence is published.
HCATonsilData has been designed to account for that. We timestamp annotations
and plan to include additional annotations if users propose them and we
validate them.
}
