% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors_20230508}
\alias{colors_20230508}
\title{Color palettes}
\format{
A list with the HEX color codes for each cell type
}
\source{
This file was created with the script in
"inst/scripts/make-annotation-colors.R
}
\description{
Color palettes used in the paper for all cell types
}
