% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{donor_metadata}
\alias{donor_metadata}
\title{Donor metadata}
\format{
A dataframe with 17 observations and 8 variables
}
\source{
Check the tonsil atlas publication
}
\description{
Data frame that contains all metadata information for the 17 donors included
in the tonsil atlas.
}
