% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCATonsilData.R
\name{HCATonsilDataInfo}
\alias{HCATonsilDataInfo}
\title{Information on the Tonsil Data}
\usage{
HCATonsilDataInfo(
  assayType = c("RNA", "ATAC", "CITE", "Spatial"),
  cellType = listCellTypes(assayType = assayType)
)
}
\arguments{
\item{assayType}{One of 'RNA', 'ATAC', 'Multiome', 'CITE' or 'Spatial'.}

\item{cellType}{A character vector of length 1 with the desired cell type.
A list of available cell types can be obtained using
\code{listCellTypes(assay_type)}.}
}
\value{
An info message is generated, and its content is returned invisibly.
}
\description{
Get information on the individual datasets of the Tonsil Atlas data
}
\examples{
HCATonsilDataInfo(assayType = "RNA", cellType = "epithelial")
}
\author{
Federico Marini
}
