% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listHCATonsilDataCellTypes.R
\name{listHCATonsilDataCellTypes}
\alias{listHCATonsilDataCellTypes}
\title{List available cell types for a dataset of the tonsil cell atlas}
\usage{
listHCATonsilDataCellTypes(dataset)
}
\arguments{
\item{dataset}{Either 'RNA', 'ATAC', 'CITE', or 'Spatial'}
}
\value{
A character vector with the available cell types for the indicated
dataset.
}
\description{
List available cell types for a dataset of the tonsil cell atlas
}
\examples{
listHCATonsilDataCellTypes(dataset = "RNA")

}
