% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{TonsilData_cellinfo_html}
\alias{TonsilData_cellinfo_html}
\title{TonsilData_cellinfo_html}
\usage{
TonsilData_cellinfo_html(
  cell_type = NULL,
  display_plot = FALSE,
  output_to = c("single_page", "html_to_embed")
)
}
\arguments{
\item{cell_type}{String character, used to define the cell type for which the
information will be displayed. Defaults to \code{NULL} - if left unspecified, the
function returns a list of the possible options}

\item{display_plot}{Logical value, defines whether to include or not a plot
for the UMAP of the selected cell type}

\item{output_to}{Character value, one of "single_page" or "html_to_embed".
Defines in which form the information is returned, either as an individual page
or simply as HTML code to directly embed into other documents.}
}
\description{
TonsilData_cellinfo_html
}
\examples{
TonsilData_cellinfo_html("PDC")
TonsilData_cellinfo_html("preB")
TonsilData_cellinfo_html("preT")
TonsilData_cellinfo_html("preT", output_to = "html_to_embed")
}
