% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{TonsilData_glossary}
\alias{TonsilData_glossary}
\title{TonsilData_glossary}
\usage{
TonsilData_glossary()
}
\value{
Data frame containing the info on the cell types included in the
TonsilDataAtlas
}
\description{
Convenience function to read directly in the file provided as \code{extdata}
}
\examples{
glossary_df <- TonsilData_glossary()
head(glossary_df)
}
