% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCATonsilData.R
\name{HCATonsilData}
\alias{HCATonsilData}
\title{Access the Tonsil Atlas data (RNA, ATAC, CITE, Spatial)}
\usage{
HCATonsilData(assayType = "RNA", cellType = "All", processedCounts = TRUE)
}
\arguments{
\item{assayType}{One of "RNA", "ATAC", "CITE" or "Spatial".}

\item{cellType}{A character vector with the cell type available.
A list of available cell types can be obtained using
\code{listCellTypes(assay_type)}.}

\item{processedCounts}{Logical scalar. If \code{TRUE}, include the processed
counts in addition to the raw counts in the SingleCellExperiment object.}
}
\value{
A named list of
\linkS4class{SingleCellExperiment} objects (one per cell_type requested
via \code{cell_type})
}
\description{
The data was downloaded from Zenodo
https://zenodo.org/record/6340174#.YkwQwX9BxH4
}
\examples{
# TODO
}
\author{
Ramon Massoni-Badosa
}
