% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCATonsilData.R
\name{HCATonsilDataInfo}
\alias{HCATonsilDataInfo}
\title{Information on the Tonsil Data}
\usage{
HCATonsilDataInfo(assayType = "RNA", cellType = "All")
}
\arguments{
\item{assayType}{One of "RNA", "ATAC", "CITE" or "Spatial".}

\item{cellType}{A character vector with the cell type available.
A list of available cell types can be obtained using
\code{listCellTypes(assay_type)}.}
}
\value{
An info message is generated, and its content is returned invisibly.
}
\description{
Get information on the individual datasets of the Tonsil Atlas data
}
\examples{
HCATonsilDataInfo(assayType = "RNA", cellType = "epithelial")
}
\author{
Federico Marini
}
