% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{TonsilData_cellinfo}
\alias{TonsilData_cellinfo}
\title{TonsilData_cellinfo}
\usage{
TonsilData_cellinfo(cell_type = NULL)
}
\arguments{
\item{cell_type}{String character, used to define the cell type for which the
information will be displayed. Defaults to \code{NULL} - if left unspecified, the
function returns a list of the possible options}
}
\value{
Invisible \code{NULL} - the information is displayed as a message in the
console.
}
\description{
TonsilData_cellinfo
}
\examples{
TonsilData_cellinfo()

TonsilData_cellinfo("PDC")
}
